% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_helpers.R
\name{buildFamilyValue}
\alias{buildFamilyValue}
\title{Title}
\usage{
buildFamilyValue(family, link)
}
\arguments{
\item{family}{Selected distribution from Family dropdown in GUI}

\item{link}{Selected valid link function from dropdown in GUI}
}
\value{
family object for use in GAM formula.
}
\description{
Title
}
\examples{
# Gamma family using log link
buildFamilyValue("Gamma", "log")
}
