% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{buildModalDialog}
\alias{buildModalDialog}
\title{buildModalDialog}
\usage{
buildModalDialog(mtitle, mdpath)
}
\arguments{
\item{mtitle}{Title of the pop-up}

\item{mdpath}{Name of markdown file located in markdown/ to display.}
}
\value{
Modal pop-up for Shiny with markdown page as content
}
\description{
buildModalDialog
}
\examples{
# Below is example of a help button observer which shows the modal dialog on click.
\donttest{
if (interactive()){
  buildModalDialog("Load Data Help", "load_help.md")
}
}

}
