% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_helpers.R
\name{buildSmooth}
\alias{buildSmooth}
\title{Title buildSmooth}
\usage{
buildSmooth(smooth_term, covars, nknots, penalised, smooth_class, by)
}
\arguments{
\item{smooth_term}{\itemize{
\item one of ("s","te","ti","t2"). Defines what smooth term is being used. See mgcv doco for formula.gam
}}

\item{covars}{\itemize{
\item List of covariates to be included in the smooth
}}

\item{nknots}{\itemize{
\item Number of knots for the smooth
}}

\item{penalised}{\itemize{
\item LOGICAL - Whether the spline will use a fixed d.f. regression spline or penalised regression spine (default)
}}

\item{smooth_class}{\itemize{
\item Which smooth class is being used for the smooth. See smooth.terms in mgcv doco
}}

\item{by}{\itemize{
\item optional - by in the spline definition. 'Default' is the default GUI input which will be ignored.
}}
}
\value{
mgcv smooth object to be passed into a gam() formula for building a model
}
\description{
Title buildSmooth
}
