% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{engine}
\alias{engine}
\title{Engine wear versus size data from gamair (Wood, 2007)}
\format{
A data frame with 2 columns and 19 rows. Each row refers to one engine model. The columns are:
\describe{
\item{wear}{an index of engine wear rate.}
\item{size}{cylinder capacity in litres.}
}
}
\source{
{gamair} Originally from... http://www3.bc.sympatico.ca/Volvo_Books/engine3.html
}
\description{
Data on engine wear against engine size for 19 Volvo car engines.
}
