% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{generateBuildNavsetCardList}
\alias{generateBuildNavsetCardList}
\title{generateBuildNavsetCardList}
\usage{
generateBuildNavsetCardList()
}
\value{
navset_card_tab for the model terms in the 'Build' module.
}
\description{
Builds the Card with tabs for model terms in the 'Build' module.
}
