% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{generateBuildTabPanel}
\alias{generateBuildTabPanel}
\title{generateBuildTabPanel}
\usage{
generateBuildTabPanel()
}
\value{
navset_card_tab for the model terms in the 'Build' module.
}
\description{
Builds the UI for the Build Module
}
