% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{generateComparisonsTabPanel}
\alias{generateComparisonsTabPanel}
\title{generateComparisonsTabPanel}
\usage{
generateComparisonsTabPanel()
}
\value{
navset_card_tab for the model terms in the 'Build' module.
}
\description{
Generates the UI for the 'Comparisons' tab in the Interpret menu.
}
