% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{generateInterpretTabPanel}
\alias{generateInterpretTabPanel}
\title{generateInterpretTabPanel}
\usage{
generateInterpretTabPanel()
}
\value{
navset_card_tab for the model terms in the 'Build' module.
}
\description{
generateInterpretTabPanel
}
