% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{generateRefTabPanel}
\alias{generateRefTabPanel}
\title{generateRefTabPanel}
\usage{
generateRefTabPanel()
}
\value{
Shiny Tab Panel containing the reference list read from markdown.
}
\description{
generateRefTabPanel
}
