% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teach_ui.R
\name{generateTeachBuildPage}
\alias{generateTeachBuildPage}
\title{generateTeachBuildPage}
\usage{
generateTeachBuildPage()
}
\value{
Accordion which contains teach section markdown files
}
\description{
generateTeachBuildPage
}
