% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teach_ui.R
\name{generateTeachGAMPage}
\alias{generateTeachGAMPage}
\title{generateTeachGAMPage}
\usage{
generateTeachGAMPage()
}
\value{
Accordion which contains teach section markdown files
}
\description{
Generates the UI for the GAM section of the Teach module.
}
