% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teach_ui.R
\name{generateTeachGLMPage}
\alias{generateTeachGLMPage}
\title{generateTeachGLMPage}
\usage{
generateTeachGLMPage()
}
\value{
Accordion which contains the GLM section of the Teach module
}
\description{
generateTeachGLMPage
}
