% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teach_ui.R
\name{generateTeachTabPanel}
\alias{generateTeachTabPanel}
\title{generateTeachTabPanel}
\usage{
generateTeachTabPanel()
}
\value{
Accordion which contains teach section markdown files
}
\description{
Generates the Tab Panel for the Learn module.
}
