% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_helpers.R
\name{getFactorsForModel}
\alias{getFactorsForModel}
\title{getFactorsForModel}
\usage{
getFactorsForModel(model)
}
\arguments{
\item{model}{mgcv model to get factor terms for}
}
\value{
char vector containing factor terms for the model
}
\description{
getFactorsForModel
}
