% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_helpers.R
\name{getGamFamilies}
\alias{getGamFamilies}
\title{getGamFamilies}
\usage{
getGamFamilies()
}
\value{
Valid subset of families supported by mgcv offered by the GUI
}
\description{
Returns a list of valid GAM families that can be used in the application.
This is those supported by mgcv with some exclusions.
}
