% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_helpers.R
\name{getGamMethods}
\alias{getGamMethods}
\title{Title getGamMethods}
\usage{
getGamMethods(family)
}
\arguments{
\item{family}{Family for distribution provided as a string}
}
\value{
Valid methods for the given family
}
\description{
Gets valid smooth parameter estimator for a given distribution family
Some that are not explicitly supported are still in the list.
}
