% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{getPlotSlopes}
\alias{getPlotSlopes}
\title{getPlotSlopes}
\usage{
getPlotSlopes(model, var_int, cond, by, type = "response", rug = FALSE)
}
\arguments{
\item{model}{userModel() from Shiny}

\item{var_int}{Variable of interest, can be a c() or single item}

\item{cond}{Conditional variable - mutually exclusive with by}

\item{by}{Marginal predictions - mutually exclusive with cond}

\item{type}{Whether to plot on link or response scale}

\item{rug}{Whether to show tick marks on the axes}
}
\value{
marginaleffects plot_slopes function
}
\description{
Takes in UI inputs from the Shiny frontend to produce plots.
Checks if either the condition or by UI element is empty and plots accordingly
Can't plot both conditional and marginal slopes at the same time,
so this function acts as a wrapper.
}
