% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_helpers.R
\name{getValidCompareArguments}
\alias{getValidCompareArguments}
\title{getValidCompareArguments}
\usage{
getValidCompareArguments()
}
\value{
Valid functions for the comparisons argument in plot_comparisons. Some have been skipped for brevity.
}
\description{
getValidCompareArguments
}
