% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_helpers.R
\name{getValidDataGridMethods}
\alias{getValidDataGridMethods}
\title{Title getValidDataGridMethods}
\usage{
getValidDataGridMethods(type)
}
\arguments{
\item{type}{String - dtype to generate methods for}
}
\value{
String - Appropriate data generation method for given dtype. Currently only factor and numeric supported.
}
\description{
Based on the dtype for a given variable, offer different generation methods for the datagrid.
Note that the defaults are mean and mode, which are not provided here.
}
