% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_helpers.R
\name{getValidDataGridTypes}
\alias{getValidDataGridTypes}
\title{getValidDataGridTypes}
\usage{
getValidDataGridTypes()
}
\value{
the three valid arguments for datagrid() grid_type argument
}
\description{
getValidDataGridTypes
}
