% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_helpers.R
\name{getValidLinkFunction}
\alias{getValidLinkFunction}
\title{Title getValidLinkFunctions}
\usage{
getValidLinkFunction(family)
}
\arguments{
\item{family}{Distribution family to get the valid link functions for}
}
\value{
Valid link functions for a given family with the first item being the canonical link function.
}
\description{
Title getValidLinkFunctions
}
