% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_helpers.R
\name{getValidSlopeFunctions}
\alias{getValidSlopeFunctions}
\title{getValidSlopeFunction}
\usage{
getValidSlopeFunctions()
}
\value{
Valid slopes as per marginaleffects package.
}
\description{
getValidSlopeFunction
}
