% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{loadMarkdown}
\alias{loadMarkdown}
\title{Load a Markdown File from the Package's inst/markdown Directory}
\usage{
loadMarkdown(filename)
}
\arguments{
\item{filename}{Name of the markdown file, e.g. "teach_intro_text.md"}
}
\value{
Shiny generation of Markdown file
}
\description{
Load a Markdown File from the Package's inst/markdown Directory
}
