% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_effects.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{conditional_effects}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{brms}{\code{\link[brms:conditional_effects.brmsfit]{conditional_effects}}}
}}

