% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_server.R
\name{studentServer}
\alias{studentServer}
\title{Title}
\usage{
studentServer(id, page_len)
}
\arguments{
\item{id}{ID to be used for server namespace}

\item{page_len}{number of pages in the student journey wizard}
}
\value{
module session holding objects and logic for teach section
}
\description{
Title
}
\examples{
# Load a student server with ID with ID "student" and 5 pages for the wizard.
# Do not test as must be run in a Shiny session.
\donttest{
if (interactive()) {
  studentServer("student", 5)
}
}
}
