% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teach_ui.R
\name{generateTeachInterpretPage}
\alias{generateTeachInterpretPage}
\title{generateTeachBuildPage}
\usage{
generateTeachInterpretPage()
}
\value{
Accordion which contains teach section markdown files
}
\description{
generateTeachBuildPage
}
